+++ CHANGES.md
@@ -5,6 +5,7 @@
   * Bugfix: Faster TickStore querying for multiple symbols simultaneously
   * Bugfix: TickStore.read now respects `allow_secondary=True`
   * Bugfix: #147 Add get_info method to ChunkStore
+  * Bugfix: Periodically re-cache the library.quota to pick up any changes
 
 ### 1.25 (2016-05-23)